requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.34
psycopg2-binary==2.9.9
python-dotenv==1.0.1
gunicorn==23.0.0
matplotlib==3.9.2
reportlab==4.2.5
